ARROW-7640: [C++][Dataset][Parquet] Detect missing compression support
The following patch ensures that reading a dataset with unspported
parquet files does not crash. It also detects the missing files so it
can still work with other files.

If any of the ColumnChunk compression is not supported, the whole file
is ignored.

Also fix an uncaught exception in parquet::arrow::reader::GetReader.